Big Improvement.

+++ common/src/main/java/me/dantero/tunnelgame/common/game/configuration/LevelConfiguration.java
@@ -7,7 +7,7 @@ import java.util.Map;
 /**
  * @author Furkan Doğan
  */
-public record LevelConfiguration(Map<Integer, Long> levelLength, Map<Integer, List<ModifiedEntitySetting>> levelEntities) {
+public record LevelConfiguration(Map<Integer, Integer> levelLength, Map<Integer, List<ModifiedEntitySetting>> levelEntities) {
 
   @Override
   public Map<Integer, List<ModifiedEntitySetting>> levelEntities() {
