+++ common/src/main/java/me/dantero/tunnelgame/common/game/configuration/LevelConfigurationBuilder.java
@@ -8,16 +8,16 @@ import java.util.Map;
 public class LevelConfigurationBuilder {
 
   private final Map<Integer, List<ModifiedEntitySetting>> levelEntities = new HashMap<>();
-  private final Map<Integer, Long> levelLength = new HashMap<>();
+  private final Map<Integer, Integer> levelLength = new HashMap<>();
 
-  public LevelConfigurationBuilder addLevelEntity(Integer level, ModifiedEntitySetting modifiedEntitySetting) {
+  public LevelConfigurationBuilder addLevelEntity(int level, ModifiedEntitySetting modifiedEntitySetting) {
     List<ModifiedEntitySetting> list = this.levelEntities.getOrDefault(level, new ArrayList<>());
     list.add(modifiedEntitySetting);
     this.levelEntities.put(level, list);
     return this;
   }
 
-  public LevelConfigurationBuilder addLevelLength(Integer level, long length) {
+  public LevelConfigurationBuilder addLevelLength(int level, int length) {
     this.levelLength.put(level, length);
     return this;
   }
